Slider - wrap in asBaseComponent and fix typings
