+++ src/components/slider/GradientSlider.tsx
@@ -1,63 +1,17 @@
 import _ from 'lodash';
 import tinycolor from 'tinycolor2';
 import React, {useCallback, useEffect, useState, useRef, useMemo} from 'react';
-import {StyleProp, ViewStyle} from 'react-native';
-import {Colors} from '../../style';
 import {asBaseComponent, forwardRef, ForwardRefInjectedProps} from '../../commons/new';
 import {ComponentStatics} from '../../typings/common';
-import Slider, {SliderProps} from './index';
+import {Colors} from '../../style';
 import {Slider as NewSlider} from '../../incubator';
+import Gradient from '../gradient';
+import {GradientSliderProps, GradientSliderTypes} from './types';
+import Slider from './index';
 import {SliderContextProps} from './context/SliderContext';
 import asSliderGroupChild from './context/asSliderGroupChild';
-import Gradient from '../gradient';
-
-type SliderOnValueChange = (value: string, alfa: number) => void;
-
-export enum GradientSliderTypes {
-  DEFAULT = 'default',
-  HUE = 'hue',
-  LIGHTNESS = 'lightness',
-  SATURATION = 'saturation'
-}
-
-export type GradientSliderProps = Omit<
-  SliderProps,
-  'onValueChange' | 'value' | 'minimumValue' | 'maximumValue' | 'step' | 'thumbHitSlop' | 'useGap'
-> & {
-  /**
-   * The gradient color
-   */
-  color?: string;
-  /**
-   * The gradient type (default, hue, lightness, saturation)
-   */
-  type?: GradientSliderTypes | `${GradientSliderTypes}`;
-  /**
-   * The gradient steps
-   */
-  gradientSteps?: number;
-  /**
-   * Callback for onValueChange, returns the updated color
-   */
-  onValueChange?: SliderOnValueChange;
-  /**
-   * If true the component will have accessibility features enabled
-   */
-  accessible?: boolean;
-  /**
-   * The container style
-   */
-  containerStyle?: StyleProp<ViewStyle>;
-  /**
-   * If true the Slider will be disabled and will appear in disabled color
-   */
-  disabled?: boolean;
-} & Partial<Pick<SliderProps, 'value' | 'minimumValue' | 'maximumValue' | 'step' | 'thumbHitSlop' | 'useGap'>>; // Fixes typing errors with the old slider.
 
 type GradientSliderComponentProps = {
-  /**
-   * Context of the slider group
-   */
   sliderContext: SliderContextProps;
 } & GradientSliderProps;
 
@@ -215,7 +169,5 @@ const GradientSlider = (props: Props) => {
 
 GradientSlider.displayName = 'GradientSlider';
 GradientSlider.types = GradientSliderTypes;
-
-// eslint-disable-next-line max-len
 // @ts-expect-error
 export default asBaseComponent<GradientSliderProps, ComponentStatics<typeof GradientSlider>>(forwardRef(asSliderGroupChild(forwardRef(GradientSlider))));
