+++ src/incubator/Slider/index.tsx
@@ -1,12 +1,15 @@
 import _ from 'lodash';
-import React, {useImperativeHandle, useCallback, useMemo, useEffect, ReactElement} from 'react';
+import React, {useImperativeHandle, useCallback, useMemo, useEffect} from 'react';
 import {StyleSheet, AccessibilityRole, StyleProp, ViewStyle, GestureResponderEvent, LayoutChangeEvent, ViewProps, AccessibilityProps} from 'react-native';
 import {useSharedValue, useAnimatedStyle, runOnJS, useAnimatedReaction, withTiming} from 'react-native-reanimated';
 import {forwardRef, ForwardRefInjectedProps, Constants} from '../../commons/new';
 import {extractAccessibilityProps} from '../../commons/modifiers';
 import {Colors, Spacings} from '../../style';
 import {StyleUtils} from 'utils';
+import {useThemeProps} from '../../hooks';
 import View from '../../components/view';
+import {ComponentStatics} from '../../typings/common';
+import {SliderProps as BaseSliderProps} from '../../components/slider/types';
 import {
   validateValues,
   getOffsetForValue,
@@ -15,68 +18,9 @@ import {
 } from './SliderPresenter';
 import Thumb from './Thumb';
 import Track from './Track';
-import {ComponentStatics} from '../../typings/common';
-export interface SliderProps extends AccessibilityProps {
-  /**
-   * Initial value
-   */
-  value?: number;
-  /**
-   * Track minimum value
-   */
-  minimumValue?: number;
-  /**
-   * Track maximum value
-   */
-  maximumValue?: number;
-  /**
-   * Initial minimum value (when useRange is true)
-   */
-  initialMinimumValue?: number;
-  /**
-   * Initial maximum value (when useRange is true)
-   */
-  initialMaximumValue?: number;
-  /**
-   * Step value of the slider. The value should be between 0 and (maximumValue - minimumValue)
-   */
-  step?: number;
-  /**
-   * The color used for the track from minimum value to current value
-   */
-  minimumTrackTintColor?: string;
-  /**
-   * The track color
-   */
-  maximumTrackTintColor?: string;
-  /**
-   * Custom render instead of rendering the track
-   */
-  renderTrack?: () => ReactElement | ReactElement[];
-  /**
-   * Callback for onValueChange
-   */
-  onValueChange?: (value: number) => void;
-  /**
-   * Callback that notifies about slider seeking is started
-   */
-  onSeekStart?: () => void;
-  /**
-   * Callback that notifies about slider seeking is finished
-   */
-  onSeekEnd?: () => void;
-  /**
-   * Callback that notifies when the reset function was invoked 
-   */
-  onReset?: () => void;
-  /**
-   * The container style
-   */
-  containerStyle?: StyleProp<ViewStyle>;
-  /**
-   * The track style
-   */
-  trackStyle?: StyleProp<ViewStyle>;
+
+// TODO: after removing old Slider move these extra props to the SliderProps in types.ts
+export interface SliderProps extends BaseSliderProps, AccessibilityProps {
   /**
    * The thumb style
    */
@@ -101,38 +45,6 @@ export interface SliderProps extends AccessibilityProps {
    * Thumb color
    */
   thumbTintColor?: string;
-  /**
-   * If true the Slider will be disabled and will appear in disabled color
-   */
-  disabled?: boolean;
-  /**
-   * If true the Slider will display a second thumb for the min value
-   */
-  useRange?: boolean;
-  /**
-   * If true the min and max thumbs will not overlap
-   */
-  useGap?: boolean;
-  /**
-   * Callback for onRangeChange. Returns values object with the min and max values
-   */
-  onRangeChange?: (values: {min: number, max: number}) => void;
-  /**
-   * If true the Slider will stay in LTR mode even if the app is on RTL mode
-   */
-  disableRTL?: boolean;
-  /**
-   * If true the component will have accessibility features enabled
-   */
-  accessible?: boolean;
-  /**
-   * The slider's test identifier
-   */
-  testID?: string;
-  /** 
-   * Whether to use the new Slider implementation using Reanimated
-   */
-  migrate?: boolean;
 }
 
 type Props = SliderProps & ForwardRefInjectedProps<SliderRef>;
@@ -151,6 +63,7 @@ const SHADOW_RADIUS = 4;
 const GAP = Spacings.s2;
 
 const Slider = React.memo((props: Props) => {
+  const themeProps = useThemeProps(props, 'IncubatorSlider');
   const {
     forwardedRef,
     useRange,
@@ -181,7 +94,7 @@ const Slider = React.memo((props: Props) => {
     accessible = true,
     testID,
     enableThumbShadow = true
-  } = props;
+  } = themeProps;
 
   const accessibilityProps = useMemo(() => {
     if (accessible) {
